既存メモ取得の処理をカスタムフックとして切り出し
